Update: Docs

+++ src/modules/appointments/appointments.controller.ts
@@ -1,7 +1,7 @@
 import { BadRequestException, Body, Controller, Delete, ForbiddenException, Get, HttpStatus, Param, ParseIntPipe, Patch, Post, Put, Query, Req, UseGuards } from '@nestjs/common';
 import { AppointmentsService } from './appointments.service';
 import { JwtAuthGuard, Roles, RolesGuard } from 'src/common';
-import { ApiBearerAuth, ApiBody, ApiForbiddenResponse, ApiOperation, ApiTags } from '@nestjs/swagger';
+import { ApiBadRequestResponse, ApiBearerAuth, ApiBody, ApiConflictResponse, ApiForbiddenResponse, ApiNotFoundResponse, ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
 import { AppointmentStatusDTO, CreateAppointmentDTO, QueryAppointmentDTO, UpdateAppointmentDTO } from './dto';
 import { Appointment } from './entity';
 
@@ -30,6 +30,7 @@ export class AppointmentsController {
         summary: 'Search appointments',
         description: 'Search appointments in the database * Requires logged in *'
     })
+    @ApiResponse({ status: 200, description: 'Return list of appointments' })
     @ApiForbiddenResponse({ description: 'Not allowed! - Customer not allowed to find others appointment' })
     async findBy(
         @Req() req: any,
@@ -50,6 +51,7 @@ export class AppointmentsController {
         summary: 'Get an appointment',
         description: 'Get an appointment from the database * Requires Admin Role *'
     })
+    @ApiResponse({ status: 200, description: 'Return an appointment' })
     async findOne(
         @Param('id', new ParseIntPipe()) id: number
     ): Promise<any> {
@@ -62,6 +64,9 @@ export class AppointmentsController {
         summary: 'Create an appointment',
         description: 'Create an appointment in the database * Requires logged in *'
     })
+    @ApiResponse({ status: 201, description: 'Return the created appointment' })
+    @ApiBadRequestResponse({ description: 'Not valid to create appointment (Customer user_id must be the same on appointment)' })
+    @ApiConflictResponse({ description: 'Another appointment already exists at this time' })
     async create(@Req() req: any, @Body() appointment: CreateAppointmentDTO): Promise<any> {
         if (req.user.role === 'customer' && req.user.id !== appointment.user_id) {
             throw new BadRequestException("Not valid to create appointment")
@@ -75,6 +80,10 @@ export class AppointmentsController {
         summary: 'Update an appointment',
         description: 'Update an appointment in the database * Requires logged in *'
     })
+    @ApiResponse({ status: 200, description: 'Return the updated appointment' })
+    @ApiBadRequestResponse({ description: 'Id not provided' })
+    @ApiForbiddenResponse({ description: 'Forbidden request' })
+    @ApiNotFoundResponse({ description: 'Appointment not found' })
     async update(
         @Param('id', new ParseIntPipe()) id: number, 
         @Req() req: any,
@@ -97,6 +106,10 @@ export class AppointmentsController {
         summary: 'Update an appointment status',
         description: 'Update an appointment status in the database * Requires Admin Role *'
     })
+    @ApiResponse({ status: 200, description: 'Return the updated appointment' })
+    @ApiBadRequestResponse({ description: 'Id not provided' })
+    @ApiNotFoundResponse({ description: 'Appointment not found' })
+    @ApiBadRequestResponse({ description: 'Invalid status' })
     async patch(
         @Param('id', new ParseIntPipe()) id: number, 
         @Body() appointmentStatusDTO: AppointmentStatusDTO
@@ -113,9 +126,18 @@ export class AppointmentsController {
         summary: 'Delete an appointment',
         description: 'Delete an appointment from the database * Requires Admin Role *'
     })
+    @ApiResponse({ status: 200, description: 'Return message of success' })
+    @ApiNotFoundResponse({ description: 'Appointment not found' })
+    @ApiBadRequestResponse({ description: 'ID is required' })
     async remove(
         @Param('id', new ParseIntPipe()) id: number
     ): Promise<any> {
-        return await this.appointmentsService.remove(id);
+        if (!id) throw new BadRequestException('ID is required');
+        const appointment = await this.appointmentsService.findOne(id);
+        if (!appointment) {
+            throw new BadRequestException('Appointment not found');
+        }
+        await this.appointmentsService.remove(id);
+        return { message: 'Appointment deleted successfully' };
     }
 }
